Name the Config fields for what they bound
MaxSize said neither "connections" nor "per server", and the client opens
a pool of that size against every server: ten servers at MaxSize 10 is a
hundred sockets, not ten. Timeout sat next to ConnectTimeout with no hint
that it bounds one operation's I/O, and IdleConnCheckThreshold named a
threshold rather than the idle duration it measures.

  MaxSize int32          -> MaxConnsPerServer int
  Timeout                -> OperationTimeout
  ConnectTimeout         -> DialTimeout
  IdleConnCheckThreshold -> IdleConnCheckAfter

The int32 was puddle's shape leaking into the public API; it is now
converted at the pool boundary.

Zero meant three different things across the struct with no stated rule.
The behaviour is unchanged, but Config now says once that the zero value
selects the field's documented default, and every field ends on "The
default is X." — including "no limit" for MaxConnLifetime and
MaxConnIdleTime. IdleConnCheckAfter keeps negative-disables as the one
documented exception, since an enable bool would default to off.

The defaults are exported to match BreakerConfig, which already exports
its own; the field docs pointed at unexported names that rendered as dead
text in godoc.

+++ pool_benchmark_test.go
@@ -11,7 +11,7 @@ import (
 
 // newBenchPool builds a pool whose constructor counts the connections it
 // creates, so benchmarks can assert creation/reuse behavior.
-func newBenchPool(b *testing.B, maxSize int32, created *atomic.Uint32) connPool {
+func newBenchPool(b *testing.B, maxSize int, created *atomic.Uint32) connPool {
 	b.Helper()
 	pool, err := newPuddlePool(func(ctx context.Context) (*Connection, error) {
 		created.Add(1)
@@ -70,7 +70,7 @@ func BenchmarkPool_Concurrent(b *testing.B) {
 	ctx := context.Background()
 
 	// RunParallel will respect GOMAXPROCS, so we use that to determine pool size
-	maxSize := int32(runtime.NumCPU())
+	maxSize := runtime.NumCPU()
 
 	var created atomic.Uint32
 	pool := newBenchPool(b, maxSize, &created)
